requirements-dev.txt
-r requirements.txt
https://github.com/nycdb/nycdb/archive/cacc0219517e557eeed61bc105278c0575e83ad0.zip#subdirectory=src
mypy==0.782
PyYAML>=5.1
pytest==6.2.5
pytest-django==4.2.0
python-dotenv==0.9.1
flake8==3.8.3
networkx==2.5.1
freezegun==1.1.0
black==22.3.0
python-geosupport==1.0.8